+++ packages/mediapipe-task-text/lib/mediapipe_flutter_text.dart
@@ -6,6 +6,7 @@
 library;
 
 export 'embedding_gemma.dart';
+export 'text_proofreader.dart';
 
 export 'universal_mediapipe_flutter_text.dart'
     if (dart.library.html) 'src/web/mediapipe_flutter_text.dart'