+++ packages/mediapipe-task-text/lib/src/io/mediapipe_flutter_text.dart
@@ -5,3 +5,5 @@
 export 'tasks/tasks.dart';
 export '../interface/embedding_gemma_types.dart';
 export 'embedding_gemma.dart';
+export '../interface/text_proofreader_types.dart';
+export 'text_proofreader.dart';